Added Requestbody for ordering pizzas. Working on errorr handling ordering/api Added in the option to view if it's a bad request or server error.

+++ PizzaPlace/Models/PizzaAmount.cs
@@ -2,4 +2,4 @@
 
 namespace PizzaPlace;
 
-public record PizzaAmount(PizzaRecipeType PizzaType, ushort Amount);
+public record PizzaAmount(PizzaRecipeType PizzaType, int Amount);
